Implement clear of Driver-side metadata.
Once Driver <-> Task communication is completed, Task-side cached blocks
should also be cleared.

+++ reef-inmemory/src/main/thrift/management.thrift
@@ -1,5 +1,5 @@
 namespace java org.apache.reef.inmemory.fs.service
 
 service SurfManagementService {
-    i64 clear()
+    i32 clear()
 }
